requirements.txt
requests
python-dotenv
numpy
fastapi
aiologger[aiofiles]
pip-chill
uvicorn
gunicorn
natasha==0.10.0
yargy==0.12.0
aiohttp